maint: generate separate man sources

Approved-by: Ken Raffenetti

+++ maint/release.pl
@@ -47,23 +47,24 @@ sub check_package
         # the user specified a dir where autoconf can be found
         if (not -x "$with_autoconf/$pack") {
             print "not found\n";
-            exit;
+            return 0;
         }
     }
     if ($with_automake and ($pack eq "automake")) {
         # the user specified a dir where automake can be found
         if (not -x "$with_automake/$pack") {
             print "not found\n";
-            exit;
+            return 0;
         }
     }
     else {
         if (`which $pack` eq "") {
             print "not found\n";
-            exit;
+            return 0;
         }
     }
     print "done\n";
+    return 1;
 }
 
 sub check_autotools_version
@@ -143,12 +144,19 @@ if (!$branch || !$version) {
     usage();
 }
 
-check_package("doctext");
-check_package("txt2man");
-check_package("git");
-check_package("latex");
-check_package("autoconf");
-check_package("automake");
+my $has_doctext = check_package("doctext");
+my $has_git = check_package("git");
+my $has_latex = check_package("latex");
+my $has_autoconf = check_package("autoconf");
+my $has_automake = check_package("automake");
+if (!$has_git || !$has_autoconf || !$has_automake) {
+    die "\tFATAL: git, autoconf, and automake are required.\n";
+}
+if (!$has_doctext || !$has_latex) {
+    print("\tWARNING: doctext or latex not found. Man pages and documentations\n");
+    print("\t         will be skipped in the release package.\n\n");
+    print("\tdoctext is available on http://wgropp.cs.illinois.edu/projects/software/sowing/\n");
+}
 print("\n");
 
 ## NOTE: Different autotools versions may result in accidental ABI chanages.
@@ -218,7 +226,7 @@ print("done\n");
 print("===> Creating configure in the main codebase... ");
 chdir($expdir);
 {
-    my $cmd = "./autogen.sh";
+    my $cmd = "./autogen.sh --with-doc";
     $cmd .= " --with-autoconf=$with_autoconf" if $with_autoconf;
     $cmd .= " --with-automake=$with_automake" if $with_automake;
     run_cmd($cmd);
@@ -253,26 +261,28 @@ run_cmd("find . -name autom4te.cache | xargs rm -rf");
 print("done\n");
 
 # Get docs
-print("===> Creating secondary codebase for the docs... ");
-run_cmd("mkdir ${expdir}-build");
-chdir("${expdir}-build");
-run_cmd("${expdir}/configure --with-device=ch4:stubnm --disable-fortran --disable-cxx");
-run_cmd("(make mandoc && make htmldoc && make latexdoc)");
-print("done\n");
+if ($has_doctext and $has_latex) {
+    print("===> Creating secondary codebase for the docs... ");
+    run_cmd("mkdir ${expdir}-build");
+    chdir("${expdir}-build");
+    run_cmd("${expdir}/configure --with-device=ch4:stubnm --disable-fortran --disable-cxx");
+    run_cmd("(make mandoc && make htmldoc && make latexdoc)");
+    print("done\n");
 
-print("===> Copying docs over... ");
-run_cmd("cp -a man ${expdir}");
-run_cmd("cp -a www ${expdir}");
-run_cmd("cp -a doc/userguide/user.pdf ${expdir}/doc/userguide");
-run_cmd("cp -a doc/installguide/install.pdf ${expdir}/doc/installguide");
-print("done\n");
+    print("===> Copying docs over... ");
+    run_cmd("cp -a man ${expdir}");
+    run_cmd("cp -a www ${expdir}");
+    run_cmd("cp -a doc/userguide/user.pdf ${expdir}/doc/userguide");
+    run_cmd("cp -a doc/installguide/install.pdf ${expdir}/doc/installguide");
+    print("done\n");
 
-print("===> Creating ROMIO docs... ");
-chdir("${expdir}/src/mpi");
-chdir("romio/doc");
-run_cmd("make");
-run_cmd("rm -f users-guide.blg users-guide.toc users-guide.aux users-guide.bbl users-guide.log users-guide.dvi");
-print("done\n");
+    print("===> Creating ROMIO docs... ");
+    chdir("${expdir}/src/mpi");
+    chdir("romio/doc");
+    run_cmd("make");
+    run_cmd("rm -f users-guide.blg users-guide.toc users-guide.aux users-guide.bbl users-guide.log users-guide.dvi");
+    print("done\n");
+}
 
 # Create the main tarball
 print("===> Creating the final mpich tarball... ");
@@ -286,7 +296,7 @@ print("===> Creating the final hydra tarball... ");
 run_cmd("cp -a ${expdir}/src/pm/hydra hydra-${version}");
 run_cmd("tar -czvf hydra-${version}.tar.gz hydra-${version}");
 run_cmd("cp -a hydra-${version}.tar.gz ${root}/");
-print("done\n\n");
+print("done\n");
 
 # Create the testsuite tarball
 print("===> Creating the final mpich-testsuite tarball... ");
